Move OrderStatusChangeTrigger into own namespace
[MAILPOET-5661]
